<관리자 비즈니스 로직 구현> 회원: 회원목록, 회원 보기, 회원삭제 상품: 상품 목록, 상품 추가, 상품 삭제 구현 완료

+++ src/main/java/com/appleshopingmall/shop/product/ProductEntity.java
@@ -6,7 +6,7 @@ import java.sql.Timestamp;
 
 @Data
 public class ProductEntity {
-    private int productID;
+    private Long productID;
     private String productName;
     private long productPrice;
     private Timestamp productData;
